Update Day 1
Task #1 Turn the first letter of its first word to uppercase.
Task #2 Implement the function shorten(s, n). It should accept a string, and a number. It should return the new string that consists of maximum n characters of the original string. If you have to cut off some chars, place "..." in the end. So, shorten('this string is too long to fit our ad', 21) returns "this string is too lo...."